Changed Units.Prefix from double to integer
- Addresses https://github.com/cellml/cellml-specification/issues/18 and https://github.com/cellml/libcellml/issues/297

- Changed reference to Units.Prefix as a double to an int in the units.cpp|h files
- Changed units.i file to ignore addUnits() functions which have Units.Prefix enum in the argument list
- Removed python tests which have Units.Prefix enum
- Updated unit tests

+++ src/api/libcellml/units.h
@@ -144,12 +144,12 @@ public:
      * @overload
      *
      * @param reference The @c std::string units reference to add.
-     * @param prefix The prefix for the unit expressed as a double.
+     * @param prefix The prefix for the unit expressed as an integer.
      * @param exponent The exponent.
      * @param multiplier The multiplier.
      * @param id The optional @c std::string id to set for this unit.
      */
-    void addUnit(const std::string &reference, double prefix, double exponent,
+    void addUnit(const std::string &reference, int prefix, double exponent,
                  double multiplier=1.0, const std::string &id="");
 
     /**
@@ -221,12 +221,12 @@ public:
      * @overload
      *
      * @param standardRef The @c StandardUnit enum units reference to add.
-     * @param prefix The prefix for the unit expressed as a double.
+     * @param prefix The prefix for the unit expressed as an integer.
      * @param exponent The exponent.
      * @param multiplier The multiplier.
      * @param id The optional @c std::string id to set for this unit.
      */
-    void addUnit(StandardUnit standardRef, double prefix, double exponent,
+    void addUnit(StandardUnit standardRef, int prefix, double exponent,
                  double multiplier=1.0, const std::string &id="");
 
     /**
